refactored App.js
